# src/backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AutoPro API",
    "description": "Documentacao das rotas da API AutoPro baseada na modelagem Cliente, Veiculo, Servico, Item de Servico e Produto.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Servidor local"
    }
  ],
...
  "paths": {
...
    "/veiculos": {
      "get": {
        "summary": "Lista todos os veiculos",
        "tags": [
          "Veiculos"
        ],
        "responses": {
          "200": {
            "description": "Lista de veiculos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Veiculo"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Cria um veiculo",
        "tags": [
          "Veiculos"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VeiculoInput"
              },
              "example": {
                "placa": "ABC1D23",
                "modelo": "Corolla",
                "ano": 2020,
                "cor": "Prata",
                "quilometragem": 45000,
                "tipoVeiculo": "Sedan",
                "motorizacao": "2.0 Flex",
                "numeroChasse": "9BWZZZ377VT004251",
                "tipoCombustivel": "Flex",
                "dataUltimaRevisao": "2026-04-20",
                "idCliente": 1
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Veiculo criado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Veiculo"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/veiculos/{id}": {
      "get": {
        "summary": "Busca um veiculo por ID",
        "tags": [
          "Veiculos"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "200": {
            "description": "Veiculo encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Veiculo"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "summary": "Atualiza um veiculo",
        "tags": [
          "Veiculos"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VeiculoInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Veiculo atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Veiculo"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "summary": "Remove um veiculo",
        "tags": [
          "Veiculos"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Id"
          }
        ],
        "responses": {
          "204": {
            "description": "Veiculo removido"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
...
    "/usuarios/login": {
      "post": {
        "summary": "Realiza login de usuario",
        "tags": [
          "Usuarios"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              },
              "example": {
                "email": "admin@email.com",
                "senha": "123456"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/usuarios": {
      "get": {
        "summary": "Lista todos os usuarios",
        "tags": [
          "Usuarios"
        ],
        "responses": {
          "200": {
            "description": "Lista de Usuario",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Usuario"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Cria um usuario",
        "tags": [
          "Usuarios"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioInput"
              },
              "example": {
                "nome": "Joao Silva",
                "email": "joao@email.com",
                "perfil": "Administrador",
                "status": "Ativo",
                "senha": "123456"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario criado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/usuarios/{id}": {
      "get": {
        "summary": "Busca um usuario por ID",
        "tags": [
          "Usuarios"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "summary": "Atualiza um usuario",
        "tags": [
          "Usuarios"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Uuid"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioUpdateInput"
              },
              "example": {
                "nome": "Joao Silva",
                "email": "joao@email.com",
                "perfil": "Supervisor",
                "status": "Ativo",
                "senha": "novaSenha123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      },
      "delete": {
        "summary": "Remove um usuario",
        "tags": [
          "Usuarios"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario removido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMessage"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "Id": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "ID do registro",
        "schema": {
          "type": "integer",
          "example": 1
        }
      },
      "Uuid": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "ID UUID do registro",
        "schema": {
          "type": "string",
          "format": "uuid",
          "example": "81f72d36-c7f1-4504-9323-81fdd59be329"
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Dados invalidos ou obrigatorios ausentes",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiMessage"
            }
          }
        }
      },
      "NotFound": {
        "description": "Registro nao encontrado",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiMessage"
            }
          }
        }
      },
      "Conflict": {
        "description": "Operacao bloqueada por relacionamento ou regra de negocio",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiMessage"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Credenciais invalidas",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiMessage"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Usuario sem permissao ou inativo",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiMessage"
            }
          }
        }
      }
    },
    "schemas": {
      "ApiMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Mensagem da API"
          }
        }
      },
...
      "Veiculo": {
        "allOf": [
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "example": 1
              },
              "dataCriacao": {
                "type": "string",
                "format": "date-time",
                "example": "2026-05-03T14:23:20.266Z"
              },
              "dataAtualizacao": {
                "type": "string",
                "format": "date-time",
                "example": "2026-05-03T14:23:20.266Z"
              }
            }
          },
          {
            "$ref": "#/components/schemas/VeiculoInput"
          }
        ]
      },
      "VeiculoInput": {
        "type": "object",
        "required": [
          "placa",
          "modelo",
          "ano",
          "cor",
          "idCliente"
        ],
        "properties": {
          "placa": {
            "type": "string",
            "example": "ABC1D23"
          },
          "modelo": {
            "type": "string",
            "example": "Corolla"
          },
          "ano": {
            "type": "integer",
            "example": 2020
          },
          "cor": {
            "type": "string",
            "example": "Prata"
          },
          "quilometragem": {
            "type": "integer",
            "example": 45000,
            "default": 0
          },
          "tipoVeiculo": {
            "type": "string",
            "nullable": true,
            "example": "Sedan"
          },
          "motorizacao": {
            "type": "string",
            "nullable": true,
            "example": "2.0 Flex"
          },
          "numeroChasse": {
            "type": "string",
            "nullable": true,
            "example": "9BWZZZ377VT004251"
          },
          "tipoCombustivel": {
            "type": "string",
            "nullable": true,
            "example": "Flex"
          },
          "dataUltimaRevisao": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "example": "2026-04-20"
          },
          "idCliente": {
            "type": "integer",
            "example": 1
          }
        }
      },
...
      "Usuario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "81f72d36-c7f1-4504-9323-81fdd59be329"
          },
          "nome": {
            "type": "string",
            "example": "Joao Silva"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "joao@email.com"
          },
          "perfil": {
            "type": "string",
            "enum": [
              "Administrador",
              "Supervisor",
              "Padr?o"
            ],
            "example": "Administrador",
            "description": "Nome do perfil associado ao idPerfil"
          },
          "status": {
            "type": "string",
            "enum": [
              "Ativo",
              "Inativo"
            ],
            "example": "Ativo"
          },
          "dataCriacao": {
            "type": "string",
            "format": "date-time",
            "example": "2026-05-03T14:23:20.266Z"
          },
          "dataAtualizacao": {
            "type": "string",
            "format": "date-time",
            "example": "2026-05-03T14:23:20.266Z"
          },
          "idPerfil": {
            "type": "integer",
            "example": 1
          }
        }
      },
      "UsuarioInput": {
        "type": "object",
        "required": [
          "nome",
          "email",
          "status",
          "senha"
        ],
        "properties": {
          "nome": {
            "type": "string",
            "example": "Joao Silva"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "joao@email.com"
          },
          "perfil": {
            "type": "string",
            "enum": [
              "Administrador",
              "Supervisor",
              "Padr?o"
            ],
            "example": "Administrador",
            "description": "Nome do perfil. Alternativa a idPerfil."
          },
          "status": {
            "type": "string",
            "enum": [
              "Ativo",
              "Inativo"
            ],
            "example": "Ativo"
          },
          "senha": {
            "type": "string",
            "format": "password",
            "minLength": 6,
            "example": "123456"
          },
          "idPerfil": {
            "type": "integer",
            "example": 1
          }
        },
        "anyOf": [
          {
            "required": [
              "perfil"
            ]
          },
          {
            "required": [
              "idPerfil"
            ]
          }
        ]
      },
      "UsuarioUpdateInput": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "example": "Joao Silva"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "joao@email.com"
          },
          "perfil": {
            "type": "string",
            "enum": [
              "Administrador",
              "Supervisor",
              "Padr?o"
            ],
            "example": "Administrador",
            "description": "Nome do perfil. Alternativa a idPerfil."
          },
          "status": {
            "type": "string",
            "enum": [
              "Ativo",
              "Inativo"
            ],
            "example": "Ativo"
          },
          "senha": {
            "type": "string",
            "format": "password",
            "minLength": 6,
            "example": "123456"
          },
          "idPerfil": {
            "type": "integer",
            "example": 2
          }
        }
      },
      "LoginInput": {
        "type": "object",
        "required": [
          "email",
          "senha"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "admin@email.com"
          },
          "senha": {
            "type": "string",
            "format": "password",
            "example": "123456"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "usuario": {
            "$ref": "#/components/schemas/Usuario"
          }
        }
      },
...
    }
  }
}